Integrate user and auth routes/models

+++ routes/boards.js
@@ -1,3 +1,4 @@
+const auth = require("../middleware/auth");
 const { Router } = require("express");
 const router = Router();
 const { Board, validate } = require("../models/board");
@@ -6,22 +7,25 @@ const { Board, validate } = require("../models/board");
 const excludeVField = { __v: 0 };
 
 // Get all boards
-router.get("/", async (req, res) => {
-  const boards = await Board.find({}, excludeVField);
+router.get("/", auth, async (req, res) => {
+  const boards = await Board.find({ user_id: req.user._id }, excludeVField);
   res.send(boards);
 });
 
 // Create a new board
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
   let { error } = validate(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
+  // Add user_id to the new board
   let board = new Board({
+    user_id: req.user._id,
     name: req.body.name,
     columns: req.body.columns,
     tasks: req.body.tasks,
   });
 
+  // Save the new board to the database
   await board.save();
 
   // Exclude the __v field from the response
@@ -31,48 +35,63 @@ router.post("/", async (req, res) => {
 });
 
 // Get board By ID
-router.get("/:id", async (req, res) => {
+router.get("/:id", auth, async (req, res) => {
   const board = await Board.findById(req.params.id).select(excludeVField);
   if (!board)
     return res.status(404).send("The board with the given ID does not exist.");
+  if (req.user._id !== board.user_id)
+    return res
+      .status(403)
+      .send("The user has no authorization over this board.");
   res.send(board);
 });
 
 // Update board By ID
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
   const { error } = validate(req.body);
   if (error) {
-    console.log(req.body);
     console.error(error.details[0].message);
     return res.status(400).send(error.details[0].message);
   }
 
-  const board = await Board.findByIdAndUpdate(
-    req.params.id,
-    {
-      name: req.body.name,
-      columns: req.body.columns,
-      tasks: req.body.tasks,
-    },
-    { new: true }
-  ).select(excludeVField);
-
-  if (!board)
+  const board = await Board.findById(req.params.id).select(excludeVField);
+  if (!board) {
     return res.status(404).send("The board with the given ID does not exist.");
+  }
 
-  res.send(board);
+  // Check if the user has authorization over this board
+  if (req.user._id !== board.user_id.toString()) {
+    return res
+      .status(403)
+      .send("The user has no authorization over this board.");
+  }
+
+  board.name = req.body.name;
+  board.columns = req.body.columns;
+  board.tasks = req.body.tasks;
+
+  const updatedBoard = await board.save();
+  res.send(updatedBoard);
 });
 
 // Delete board By ID
-router.delete("/:id", async (req, res) => {
-  const board = await Board.findByIdAndDelete(req.params.id).select(
+router.delete("/:id", auth, async (req, res) => {
+  const board = await Board.findById(req.params.id).select(excludeVField);
+  if (!board) {
+    return res.status(404).send("The board with the given ID does not exist.");
+  }
+
+  // Check if the user has authorization over this board
+  if (req.user._id !== board.user_id.toString()) {
+    return res
+      .status(403)
+      .send("The user has no authorization over this board.");
+  }
+
+  const deletedBoard = await Board.findByIdAndDelete(req.params.id).select(
     excludeVField
   );
-
-  if (!board)
-    return res.status(404).send("The board with the given ID does not exist.");
-
-  res.send(board);
+  res.send(deletedBoard);
 });
 
 module.exports = router;
